requirements.txt
Adafruit-Blinka==8.56.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-framebuf==1.6.7
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-ssd1306==2.12.19
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
colorzero==2.0
gpiod==2.3.0
gpiozero==2.0.1
lgpio==0.2.2.0
mutagen==1.47.0
numpy==2.2.5
pigpio==1.78
pillow==11.2.1
pyftdi==0.56.0
pygame==2.6.1
pyserial==3.5
pyusb==1.3.1
rpi-lgpio==0.6
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
smbus2==0.5.0
spidev==3.6
sysv-ipc==1.1.0
tinytag==2.1.1
typing_extensions==4.13.2
